+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
+## Unreleased
+* Tap on any part of the video preview to pause or play the video.
 ## v0.3 (PRERELEASE)
 * Changed scale factor of avatar image in channel view.
 * Added description field